pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "diffusion_deep_dream_research"
version = "0.0.1"
description = "The goal of this project is to develop a method to generate monosemantic dDeepDream representations of neurons in diffusion models."
authors = [
  { name = "Adam Szokalski" },
]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: BSD License"
]
dependencies = [
    "accelerate",
    "datasets",
    "diffusers",
    "einops",
    "gdown",
    "graphviz",
    "hydra-core",
    "hydra-submitit-launcher",
    "imageio",
    "ipython",
    "ipywidgets",
    "jaxtyping",
    "jupyter",
    "jupyterlab",
    "kornia",
    "lightning",
    "loguru",
    "matplotlib",
    "mkdocs",
    "natsort",
    "notebook",
    "numpy",
    "onnx",
    "onnxscript",
    "opencv-python",
    "pandas",
    "pip",
    "pydantic",
    "pydantic-settings",
    "pytest",
    "python-dotenv",
    "ruff",
    "safetensors",
    "scikit-learn",
    "scipy",
    "simple-parsing",
    "submitit",
    "torch",
    "torch-lucent",
    "torchinfo",
    "torchview",
    "torchvision",
    "tqdm",
    "transformers",
    "typer",
    "torch-lucent @ git+https://github.com/aszokalski/lucent.git@custom-device-support"
]
requires-python = "~=3.12.0"

[tool.setuptools.packages.find]
where = ["."]  # Look in the root directory
# IMPORTANT: Add the name you want to use for importing the submodule here
include = ["diffusion_deep_dream_research", "submodules"]

# 2. Map the import name to the actual directory
[tool.setuptools.package-dir]
"diffusion_deep_dream_research" = "diffusion_deep_dream_research"
"submodules" = "submodules"

[tool.ruff]
line-length = 99
src = ["diffusion_deep_dream_research"]
include = ["pyproject.toml", "diffusion_deep_dream_research/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # Add import sorting

[tool.ruff.lint.isort]
known-first-party = ["diffusion_deep_dream_research"]
force-sort-within-sections = true

[tool.pyright]
reportPrivateImportUsage = "none"

[tool.ty.rules]
# Ignore unknown signature of callables. We are using third-party libraries which lack typing.
call-top-callable = "ignore" 


[tool.ty.src]
include = ["diffusion_deep_dream_research", "tests"]